moved reified constraints to invariants
