+++ include/invariantgraph/invariants/boolXorReifNode.hpp
