+++ include/invariantgraph/invariants/eqReifNode.hpp
