+++ include/invariantgraph/invariants/intLinNeReifNode.hpp
