+++ src/invariantgraph/invariants/intNeReifNode.cpp
@@ -1,19 +1,20 @@
-#include "invariantgraph/views/boolXorReifNode.hpp"
+#include "invariantgraph/invariants/intNeReifNode.hpp"
 
 #include "../parseHelper.hpp"
 #include "invariantgraph/constraints/intNeNode.hpp"
+#include "invariantgraph/invariants/reifiedConstraint.hpp"
 
-std::unique_ptr<invariantgraph::BoolXorReifNode>
-invariantgraph::BoolXorReifNode::fromModelConstraint(
+std::unique_ptr<invariantgraph::IntNeReifNode>
+invariantgraph::IntNeReifNode::fromModelConstraint(
     const fznparser::FZNModel&, const fznparser::Constraint& constraint,
     const std::function<VariableNode*(MappableValue&)>& variableMap) {
-  assert(constraint.name == "bool_xor_reif");
+  assert(constraint.name == "int_ne_reif");
   assert(constraint.arguments.size() == 3);
 
   auto a = mappedVariable(constraint.arguments[0], variableMap);
   auto b = mappedVariable(constraint.arguments[1], variableMap);
   auto r = mappedVariable(constraint.arguments[2], variableMap);
 
-  return std::make_unique<BoolXorReifNode>(std::make_unique<IntNeNode>(a, b),
-                                           r);
+  return std::make_unique<invariantgraph::IntNeReifNode>(
+      std::make_unique<invariantgraph::IntNeNode>(a, b), r);
 }
