--- src/invariantgraph/views/intNeReifNode.cpp
@@ -1,20 +1,0 @@
-#include "invariantgraph/views/intNeReifNode.hpp"
-
-#include "../parseHelper.hpp"
-#include "invariantgraph/constraints/intNeNode.hpp"
-#include "invariantgraph/views/reifiedConstraint.hpp"
-
-std::unique_ptr<invariantgraph::IntNeReifNode>
-invariantgraph::IntNeReifNode::fromModelConstraint(
-    const fznparser::FZNModel&, const fznparser::Constraint& constraint,
-    const std::function<VariableNode*(MappableValue&)>& variableMap) {
-  assert(constraint.name == "int_ne_reif");
-  assert(constraint.arguments.size() == 3);
-
-  auto a = mappedVariable(constraint.arguments[0], variableMap);
-  auto b = mappedVariable(constraint.arguments[1], variableMap);
-  auto r = mappedVariable(constraint.arguments[2], variableMap);
-
-  return std::make_unique<invariantgraph::IntNeReifNode>(
-      std::make_unique<invariantgraph::IntNeNode>(a, b), r);
-}
