+++ test/invariantgraph/invariants/tEqReifNode.cpp
@@ -1,38 +1,33 @@
 #include "../nodeTestBase.hpp"
 #include "core/propagationEngine.hpp"
-#include "invariantgraph/views/leReifNode.hpp"
+#include "invariantgraph/invariants/eqReifNode.hpp"
 
-class IntLeReifNodeTest : public NodeTestBase {
+class EqReifNodeTest : public NodeTestBase {
  public:
   INT_VARIABLE(a, 5, 10);
   INT_VARIABLE(b, 2, 7);
   INT_VARIABLE(r, 0, 1);
 
-  fznparser::Constraint constraint{"int_le_reif", {"a", "b", "r"}, {}};
+  fznparser::Constraint constraint{"int_eq_reif", {"a", "b", "r"}, {}};
 
   fznparser::FZNModel model{{}, {a, b, r}, {constraint}, fznparser::Satisfy{}};
 
-  std::unique_ptr<invariantgraph::LeReifNode> node;
+  std::unique_ptr<invariantgraph::EqReifNode> node;
 
-  IntLeReifNodeTest() : NodeTestBase(model) {}
+  EqReifNodeTest() : NodeTestBase(model) {}
 
   void SetUp() override {
-    node = makeNode<invariantgraph::LeReifNode>(constraint);
+    node = makeNode<invariantgraph::EqReifNode>(constraint);
   }
 };
 
-TEST_F(IntLeReifNodeTest, construction) {
+TEST_F(EqReifNodeTest, construction) {
   EXPECT_EQ(node->definedVariables().size(), 1);
   EXPECT_EQ(*node->definedVariables()[0]->variable(),
             invariantgraph::VariableNode::FZNVariable(r));
-
-  EXPECT_EQ(node->inputs().size(), 2);
-  for (const auto& input : node->inputs()) {
-    EXPECT_EQ(input->inputFor().size(), 1);
-  }
 }
 
-TEST_F(IntLeReifNodeTest, application) {
+TEST_F(EqReifNodeTest, application) {
   PropagationEngine engine;
   engine.open();
   registerVariables(engine, {a.name, b.name});
@@ -49,11 +44,22 @@ TEST_F(IntLeReifNodeTest, application) {
   // a, b
   EXPECT_EQ(engine.searchVariables().size(), 2);
 
-  // a, b, sum
+  // a, b and r
   EXPECT_EQ(engine.numVariables(), 3);
 }
 
-TEST_F(IntLeReifNodeTest, propagation) {
+static Int isViolating(const std::vector<Int>& values) {
+  for (size_t i = 0; i < values.size(); i++) {
+    for (size_t j = i + 1; j < values.size(); j++) {
+      if (values.at(i) != values.at(j)) {
+        return 1;
+      }
+    }
+  }
+  return 0;
+}
+
+TEST_F(EqReifNodeTest, propagation) {
   PropagationEngine engine;
   engine.open();
   registerVariables(engine, {a.name, b.name});
@@ -86,8 +92,7 @@ TEST_F(IntLeReifNodeTest, propagation) {
       engine.query(outputId);
       engine.endProbe();
 
-      EXPECT_EQ(engine.currentValue(outputId),
-                static_cast<Int>(values.at(0) > values.at(1)));
+      EXPECT_EQ(engine.currentValue(outputId) > 0, isViolating(values));
     }
   }
 }
