+++ test/invariantgraph/invariants/tIntLinLeReifNode.cpp
@@ -1,34 +1,39 @@
 #include "../nodeTestBase.hpp"
 #include "core/propagationEngine.hpp"
-#include "invariantgraph/views/intNeReifNode.hpp"
+#include "invariantgraph/invariants/intLinLeReifNode.hpp"
 
-class IntNeReifNodeTest : public NodeTestBase {
+class IntLinLeReifNodeTest : public NodeTestBase {
  public:
   INT_VARIABLE(a, 5, 10);
   INT_VARIABLE(b, 2, 7);
   INT_VARIABLE(r, 0, 1);
   Int c{3};
+  std::vector<Int> coeffs{1, -1};
 
-  fznparser::Constraint constraint{"int_ne_reif", {"a", "b", "r"}, {}};
+  fznparser::Constraint constraint{
+      "int_lin_le_reif",
+      {fznparser::Constraint::ArrayArgument{coeffs.at(0), coeffs.at(1)},
+       fznparser::Constraint::ArrayArgument{"a", "b"}, c, "r"},
+      {}};
 
   fznparser::FZNModel model{{}, {a, b, r}, {constraint}, fznparser::Satisfy{}};
 
-  std::unique_ptr<invariantgraph::IntNeReifNode> node;
+  std::unique_ptr<invariantgraph::IntLinLeReifNode> node;
 
-  IntNeReifNodeTest() : NodeTestBase(model) {}
+  IntLinLeReifNodeTest() : NodeTestBase(model) {}
 
   void SetUp() override {
-    node = makeNode<invariantgraph::IntNeReifNode>(constraint);
+    node = makeNode<invariantgraph::IntLinLeReifNode>(constraint);
   }
 };
 
-TEST_F(IntNeReifNodeTest, construction) {
+TEST_F(IntLinLeReifNodeTest, construction) {
   EXPECT_EQ(node->definedVariables().size(), 1);
   EXPECT_EQ(*node->definedVariables()[0]->variable(),
             invariantgraph::VariableNode::FZNVariable(r));
 }
 
-TEST_F(IntNeReifNodeTest, application) {
+TEST_F(IntLinLeReifNodeTest, application) {
   PropagationEngine engine;
   engine.open();
   registerVariables(engine, {a.name, b.name});
@@ -49,7 +54,16 @@ TEST_F(IntNeReifNodeTest, application) {
   EXPECT_EQ(engine.numVariables(), 3);
 }
 
-TEST_F(IntNeReifNodeTest, propagation) {
+static bool isViolating(const std::vector<Int>& coeffs,
+                        const std::vector<Int>& values, const Int expected) {
+  Int sum = 0;
+  for (size_t i = 0; i < values.size(); ++i) {
+    sum += coeffs.at(i) * values.at(i);
+  }
+  return sum > expected;
+}
+
+TEST_F(IntLinLeReifNodeTest, propagation) {
   PropagationEngine engine;
   engine.open();
   registerVariables(engine, {a.name, b.name});
@@ -57,7 +71,6 @@ TEST_F(IntNeReifNodeTest, propagation) {
   node->registerWithEngine(engine, _variableMap);
 
   std::vector<VarId> inputs;
-  EXPECT_EQ(node->inputs().size(), 2);
   for (auto* const inputVariable : node->inputs()) {
     EXPECT_TRUE(_variableMap.contains(inputVariable));
     inputs.emplace_back(_variableMap.at(inputVariable));
@@ -65,6 +78,8 @@ TEST_F(IntNeReifNodeTest, propagation) {
 
   EXPECT_TRUE(_variableMap.contains(node->definedVariables().at(0)));
   const VarId outputId = _variableMap.at(node->definedVariables().at(0));
+  EXPECT_EQ(inputs.size(), 2);
+
   std::vector<Int> values(inputs.size());
   engine.close();
 
@@ -82,8 +97,9 @@ TEST_F(IntNeReifNodeTest, propagation) {
       engine.query(outputId);
       engine.endProbe();
 
-      EXPECT_EQ(engine.currentValue(outputId),
-                static_cast<Int>(values.at(0) == values.at(1)));
+      const Int viol = engine.currentValue(outputId);
+
+      EXPECT_EQ(viol > 0, isViolating(coeffs, values, c));
     }
   }
 }
