+++ test/invariantgraph/invariants/tIntNeReifNode.cpp
@@ -1,32 +1,34 @@
 #include "../nodeTestBase.hpp"
 #include "core/propagationEngine.hpp"
-#include "invariantgraph/views/ltReifNode.hpp"
+#include "invariantgraph/invariants/intNeReifNode.hpp"
 
-class IntLtReifNodeTest : public NodeTestBase {
+class IntNeReifNodeTest : public NodeTestBase {
  public:
   INT_VARIABLE(a, 5, 10);
   INT_VARIABLE(b, 2, 7);
   INT_VARIABLE(r, 0, 1);
+  Int c{3};
 
-  fznparser::Constraint constraint{"int_lt_reif", {"a", "b", "r"}, {}};
+  fznparser::Constraint constraint{"int_ne_reif", {"a", "b", "r"}, {}};
+
   fznparser::FZNModel model{{}, {a, b, r}, {constraint}, fznparser::Satisfy{}};
 
-  std::unique_ptr<invariantgraph::LtReifNode> node;
+  std::unique_ptr<invariantgraph::IntNeReifNode> node;
 
-  IntLtReifNodeTest() : NodeTestBase(model) {}
+  IntNeReifNodeTest() : NodeTestBase(model) {}
 
   void SetUp() override {
-    node = makeNode<invariantgraph::LtReifNode>(constraint);
+    node = makeNode<invariantgraph::IntNeReifNode>(constraint);
   }
 };
 
-TEST_F(IntLtReifNodeTest, construction) {
+TEST_F(IntNeReifNodeTest, construction) {
   EXPECT_EQ(node->definedVariables().size(), 1);
   EXPECT_EQ(*node->definedVariables()[0]->variable(),
             invariantgraph::VariableNode::FZNVariable(r));
 }
 
-TEST_F(IntLtReifNodeTest, application) {
+TEST_F(IntNeReifNodeTest, application) {
   PropagationEngine engine;
   engine.open();
   registerVariables(engine, {a.name, b.name});
@@ -43,11 +45,11 @@ TEST_F(IntLtReifNodeTest, application) {
   // a, b
   EXPECT_EQ(engine.searchVariables().size(), 2);
 
-  // a, b, sum
+  // a, b and r
   EXPECT_EQ(engine.numVariables(), 3);
 }
 
-TEST_F(IntLtReifNodeTest, propagation) {
+TEST_F(IntNeReifNodeTest, propagation) {
   PropagationEngine engine;
   engine.open();
   registerVariables(engine, {a.name, b.name});
@@ -80,8 +82,8 @@ TEST_F(IntLtReifNodeTest, propagation) {
       engine.query(outputId);
       engine.endProbe();
 
-      EXPECT_EQ(engine.currentValue(outputId),
-                static_cast<Int>(values.at(0) >= values.at(1)));
+      EXPECT_EQ(engine.currentValue(outputId) > 0,
+                static_cast<Int>(values.at(0) == values.at(1)));
     }
   }
 }
